Split out Import/Export functions...
Also go over a couple of SystemFiles/Formats

Better align "nffil" message with WMA
