+++ .pep8
@@ -1,3 +1,3 @@
 [pep8]
-ignore = E501,W291
+ignore = E501,W291,W503
 exclude = magic.py,build/,.tox/
