+++ mathics/SystemFiles/Formats/ZIP/Import.wl
@@ -1,12 +1,39 @@
-(* ZIP compressed file and file archive Importer *)
+(* ::Package:: *)
+(* ZIP compressed file and file archive Importer.
+   This is used by Import[].
+ *)
 
-Begin["System`ZIP`"]
+Begin["System`Convert`CommonArchiveDump`"]
+
+
+$ZIPHiddenElements = {_String, "FileNamesLegacy"};
+
+$ZIPDocumentedElements = {"FileNames"};
+
+$ZIPAvailableElements = SortBy[Join[$ZIPHiddenElements, $ZIPDocumentedElements], ToString];
+
+GetElements[___] :=
+	"Elements" ->
+		SortBy[
+			$ZIPDocumentedElements,
+			ToString
+		];
+
 ImportExport`RegisterImport[
     "ZIP",
-    ImportZip,
+    ImportZIP,
+    {}, (* Post importer function(s) *)
     FunctionChannels -> {"Streams"},
-    AvailableElements -> {"FileNames", "Summary"},
-    DefaultElement -> "FileNames"
+    AvailableElements -> $ZIPAvailableElements,
+    BinaryFormat -> True,
+    DefaultElement -> "FileNames",
+    HiddenElements -> $ZIPHiddenElements,
+    SkipPostImport -> <|
+	    "FileNames" -> Automatic,
+	    "Summary" -> Automatic,
+	    "FileNamesLegacy" -> None, (* Disable SkipPostImport for FileNamesLegacy explicitly*)
+	    (_String | _List) -> Automatic
+	|>
 ]
 
 End[]
