+++ mathics/builtin/files_io/__init__.py
@@ -9,4 +9,4 @@ Input/Output, Files, and Filesystem
 
 
 # This tells documentation how to sort this module
-sort_order = "mathics.builtin.input-output-files-and-filesystem"
+sort_order = "mathics.builtin.filesystem"
