+++ mathics/eval/files_io/importexport.py
@@ -3,16 +3,13 @@ Functions for figuring out a filetype or MIME type a given
 file path.
 """
 
-import json
 import mimetypes
 import os.path as osp
-import zipfile
 from itertools import chain
 from typing import Dict, Final, Optional
 
 from mathics.core.atoms import ByteArray, String
 from mathics.core.builtin import get_option
-from mathics.core.convert.expression import to_mathics_list
 from mathics.core.convert.python import from_python
 from mathics.core.expression import Expression
 from mathics.core.list import ListExpression
@@ -215,6 +212,7 @@ def eval_Import(
     """
     Basic implemenation beind Import[].
     """
+
     current_predetermined_out = evaluation.predetermined_out
     # Check elements
     if elements.has_form("List", None):
@@ -408,64 +406,6 @@ def eval_Import_Elements(file_format: str, evaluation):
     # under the option "Elements".
     _, _, _, options = IMPORTERS[filetype]
     return options.get("System`AvailableElements")
-
-
-# FIXME:
-# We should not be extracting everything and returning a list of rules.
-# provide a better interface.
-def eval_JSONImport(json_path: str) -> ListExpression:
-    """Takes a ZIP file path and returns a list of file names/paths contained inside."""
-    with open(json_path, "r") as json_file:
-        json_data = json.load(json_file)
-        mathics_json = from_python(json_data)
-        exprs = [
-            Expression(
-                SymbolRule,
-                String("Data"),
-                mathics_json,
-            ),
-            Expression(
-                SymbolRule,
-                String("Dataset"),
-                mathics_json,
-            ),
-        ]
-        return ListExpression(*exprs)
-
-
-# FIXME:
-# We should not be extracting everything and returning a list of rules.
-# provide a better interface.
-def eval_ZIPImport(zip_path: str) -> ListExpression:
-    """Takes a ZIP file path and returns a list of file names/paths contained inside."""
-    with zipfile.ZipFile(zip_path, "r") as archive:
-        # FIXME: Using "filenames" for "Summary" items is not quite right.
-        filenames = archive.namelist()
-        mathics_filenames = to_mathics_list(*filenames)
-        exprs = [
-            Expression(
-                SymbolRule,
-                String("FileNames"),
-                mathics_filenames,
-            ),
-            Expression(
-                SymbolRule,
-                String("Summary"),
-                mathics_filenames,
-            ),
-        ]
-
-        if filenames:
-            for filename in filenames:
-                exprs.append(
-                    Expression(
-                        SymbolRule,
-                        String(filename),
-                        String(archive.read(filename).decode("utf-8")),
-                    )
-                )
-
-        return ListExpression(*exprs)
 
 
 def get_results(
